move code as is and remove from old registry
