+++ pkg/registry/apps/alerting/notifications/register.go
@@ -14,10 +14,10 @@ import (
 	"k8s.io/kube-openapi/pkg/spec3"
 
 	"github.com/grafana/grafana/apps/alerting/notifications/pkg/apis/resource"
-	"github.com/grafana/grafana/pkg/registry/apis/alerting/notifications/receiver"
-	"github.com/grafana/grafana/pkg/registry/apis/alerting/notifications/routingtree"
-	"github.com/grafana/grafana/pkg/registry/apis/alerting/notifications/templategroup"
-	"github.com/grafana/grafana/pkg/registry/apis/alerting/notifications/timeinterval"
+	"github.com/grafana/grafana/pkg/registry/apps/alerting/notifications/receiver"
+	"github.com/grafana/grafana/pkg/registry/apps/alerting/notifications/routingtree"
+	"github.com/grafana/grafana/pkg/registry/apps/alerting/notifications/templategroup"
+	"github.com/grafana/grafana/pkg/registry/apps/alerting/notifications/timeinterval"
 	"github.com/grafana/grafana/pkg/services/accesscontrol"
 	"github.com/grafana/grafana/pkg/services/apiserver/builder"
 	"github.com/grafana/grafana/pkg/services/apiserver/endpoints/request"
